+++ src/utilityfunc.cpp
@@ -3,16 +3,16 @@
 
 int argmax(vector<double> & x) {
   int maxind=0;
-  for(int i=1;i<x.size();i++) 
+  for(size_t i=1;i<x.size();i++) 
     if(x[i]>x[maxind]) maxind=i;
   return(maxind);
 }
 
 
 unsigned char**uc2hap() {
-  int n = 256;
+  size_t n = 256;
   unsigned char **ret = newMatrix<unsigned char>(n,8);
-  for(unsigned int i=0;i<n;i++) {
+  for(size_t i=0;i<n;i++) {
     bitset<8> bs(i);
     for(int j=0;j<8;j++) ret[i][j] = bs[j];
   }
@@ -21,9 +21,9 @@ unsigned char**uc2hap() {
 }
 
 unsigned char** buildLookup() {
-  int n = 256;
+  size_t n = 256;
   unsigned char **ret = newMatrix<unsigned char>(n,n);
-  for(unsigned int i=0;i<n;i++) {
+  for(size_t i=0;i<n;i++) {
     bitset<8> bs1(i);
     for(unsigned int j=0;j<n;j++) {
       bitset<8> bs2(j);
